% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simukde.R
\docType{package}
\name{simukde}
\alias{simukde}
\alias{simukde-package}
\title{Simulation with Kernel Density Estimation}
\description{
The simukde package provides a function which generates random values from a univariate and multivariate continuous distribution by using kernel density estimation based on a sample. The function uses Accept Reject Method.
}
\author{
MAKHGAL Ganbold, National University of Mongolia
}
